fix: per-function Z3 rollback, pure ISLE rewrites for control flow, async support
Per-function Z3 verification rollback:
- constant_folding and optimize_advanced_instructions now revert
  individual functions on Z3 counterexample instead of aborting the
  entire pass.

Pure ISLE rewrites for control-flow functions:
- Functions with BrIf/BrTable are no longer skipped entirely.
- New rewrite_pure() applies structural ISLE rules (constant folding,
  algebraic identities, strength reduction) without dataflow env tracking.
- rewrite_with_dataflow() (formerly simplify_with_env) is used only for
  straight-line code where local/memory propagation is safe.
- Backward-compatible aliases simplify() and simplify_with_env() retained.

Async component model support:
- Enable cm_async, cm_async_stackful, cm_async_builtins in wasmparser
  Validator. Components with task.return now parse correctly.

Stats percentage overflow fix:
- Cast to f64 before subtraction to prevent unsigned wraparound.

Fixes: H-19
Trace: REQ-5, REQ-11, SC-8

+++ loom-cli/src/main.rs
@@ -156,7 +156,7 @@ impl OptimizationStats {
         if before == 0 {
             0.0
         } else {
-            ((before - after) as f64 / before as f64) * 100.0
+            ((before as f64 - after as f64) / before as f64) * 100.0
         }
     }
 }
@@ -285,7 +285,7 @@ fn optimize_command(
                         stats.original_size,
                         stats.optimized_size,
                         if stats.original_size > 0 {
-                            ((stats.original_size - stats.optimized_size) as f64
+                            ((stats.original_size as f64 - stats.optimized_size as f64)
                                 / stats.original_size as f64)
                                 * 100.0
                         } else {
@@ -301,7 +301,8 @@ fn optimize_command(
                             "Module size:  {} → {} bytes ({:.1}% reduction)",
                             stats.original_module_size,
                             stats.optimized_module_size,
-                            ((stats.original_module_size - stats.optimized_module_size) as f64
+                            ((stats.original_module_size as f64
+                                - stats.optimized_module_size as f64)
                                 / stats.original_module_size as f64)
                                 * 100.0
                         );
